ac-65 add uuid to required parameter in posting a new product transaction.
The agritrader could create a farmer offline, and link a new product transaction
to that new farmer which is not in the database. If the database sends a new uuid then
the agritrader must update all the product transactions. Requiring a uuid and using that
one makes the agritrader alot easier.